Se agrego: Creación en automatica en la base de datos, solo admin puede agregar usuarios y filtro por rol

+++ usuario/src/main/java/com/TechMant/usuario/dto/RolDTO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 
 @Data
 public class Rol {
-    private Long id;       // Nombre exacto al JSON ("idRol")
+    private Integer id;       // Nombre exacto al JSON ("idRol")
     private String nombreRol;
 }
